+++ src/main/java/api/domain/pets/PetStatus.java
@@ -1,5 +1,12 @@
 package api.domain.pets;
 
-public enum PetStatus {
-    available, pending, sold;
+import java.util.Random;
+
+public enum PetStatusEnum {
+    AVAILABLE, PENDING, SOLD;
+
+    public static PetStatusEnum generatePetRandomStatus() {
+        Random randomStatus = new Random();
+        return values()[randomStatus.nextInt(values().length)];
+    }
 }
